Made scene manager with tabs
In the SceneManagerComplete scene is a example of the newly made scene manager. It allows you to click on tabs which are organised in categories. There was a bug found where the scene manager added a category in a category it didnt belong. This was quickly fixed after facing this issue.